需求[#22204]:SR-20201222-02-20 任务结果数据历史保存
    [描述说明]: 任务结果数据历史保存
    [问题原因]:
    [解决方法]:
    [测试建议]:

+++ iot-center-platform/src/main/java/com/yjh/platform/module/user/service/TCameraPresetService.java
@@ -40,10 +40,12 @@ public class TCameraPresetService {
     @Transactional(rollbackFor = Exception.class)
     public int deleteByPrimaryId(Long presetId) {
         List<Long> instanceIdList = tCameraPresetDao.selectInstanceIdList(presetId);
-        tCameraPresetDao.deleteInstance(instanceIdList);//tcpi
-        tCameraPresetDao.deletePlanInstance(instanceIdList);//tcplan
-        tCameraPresetDao.deletePointInstance(instanceIdList);//tcpattr
-        tAlgorithmConfDao.deleteByPrimaryId(presetId);//tac
+        if(instanceIdList != null && instanceIdList.size() >0){
+            tCameraPresetDao.deleteInstance(instanceIdList);//tcpi
+            tCameraPresetDao.deletePlanInstance(instanceIdList);//tcplan
+            tCameraPresetDao.deletePointInstance(instanceIdList);//tcpattr
+            tAlgorithmConfDao.deleteByPrimaryId(presetId);//tac
+        }
         return this.tCameraPresetDao.deleteByPrimaryId(presetId);
     }
